TS-31571 Moved the logging classes to a separate package
