+++ agent/src/main/java/com/teamscale/jacoco/agent/ResourceBase.java
@@ -3,7 +3,7 @@ package com.teamscale.jacoco.agent;
 import com.teamscale.client.CommitDescriptor;
 import com.teamscale.client.TeamscaleServer;
 import com.teamscale.jacoco.agent.testimpact.TestwiseCoverageAgent;
-import com.teamscale.jacoco.agent.util.LoggingUtils;
+import com.teamscale.jacoco.agent.logging.LoggingUtils;
 import com.teamscale.report.testwise.model.RevisionInfo;
 import org.conqat.lib.commons.string.StringUtils;
 import org.slf4j.Logger;
