+++ agent/src/main/java/com/teamscale/jacoco/agent/configuration/ProcessInformationRetriever.java
@@ -41,7 +41,6 @@ public class ProcessInformationRetriever {
 		}
 	}
 
-
 	/**
 	 * Returns a string that <i>probably</i> contains the PID.
 	 * <p>
