+++ agent/src/main/java/com/teamscale/jacoco/agent/testimpact/CoverageToExecFileStrategy.java
@@ -2,7 +2,7 @@ package com.teamscale.jacoco.agent.testimpact;
 
 import com.teamscale.jacoco.agent.JacocoRuntimeController;
 import com.teamscale.jacoco.agent.options.AgentOptions;
-import com.teamscale.jacoco.agent.util.LoggingUtils;
+import com.teamscale.jacoco.agent.logging.LoggingUtils;
 import com.teamscale.report.testwise.jacoco.cache.CoverageGenerationException;
 import com.teamscale.report.testwise.model.TestExecution;
 import com.teamscale.report.testwise.model.TestInfo;
