+++ agent/src/main/java/com/teamscale/jacoco/agent/upload/delay/DelayedUploader.java
@@ -12,7 +12,7 @@ import org.slf4j.Logger;
 
 import com.teamscale.jacoco.agent.upload.IUploader;
 import com.teamscale.jacoco.agent.util.DaemonThreadFactory;
-import com.teamscale.jacoco.agent.util.LoggingUtils;
+import com.teamscale.jacoco.agent.logging.LoggingUtils;
 import com.teamscale.report.jacoco.CoverageFile;
 
 /**
