+++ agent/src/main/java/com/teamscale/jacoco/agent/util/Benchmark.java
@@ -1,5 +1,6 @@
 package com.teamscale.jacoco.agent.util;
 
+import com.teamscale.jacoco.agent.logging.LoggingUtils;
 import org.slf4j.Logger;
 
 /**
