+++ agent/src/main/resources/com/teamscale/jacoco/agent/logback-default-debugging.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <configuration>
-	<define name="defaultLogDir" class="com.teamscale.jacoco.agent.util.DebugLogDirectoryPropertyDefiner" />
+	<define name="defaultLogDir" class="com.teamscale.jacoco.agent.logging.DebugLogDirectoryPropertyDefiner" />
 
 	<appender name="CONSOLE" class="shadow.ch.qos.logback.core.ConsoleAppender">
 		<encoder>
